+++ impl/stdlib/common/src/main/kotlin/com/github/jcornaz/collekt/stdlib/StdlibMap.kt
@@ -2,7 +2,7 @@ package com.github.jcornaz.collekt.stdlib
 
 import com.github.jcornaz.collekt.api.*
 
-public class StdlibMap<K, out V> private constructor(private val map: Map<K, V>) : PersistentMap<K, V> {
+public class StdlibMap<K, out V> private constructor(private val map: Map<K, V>) : AbstractPersistentMap<K, V>() {
     companion object Factory : PersistentMapFactory {
         private val empty = StdlibMap<Any?, Nothing>(emptyMap())
 
@@ -51,18 +51,6 @@ public class StdlibMap<K, out V> private constructor(private val map: Map<K, V>)
 
     override fun minus(keys: Iterable<K>): PersistentMap<K, V> =
             wrap(map - keys)
-
-    override fun equals(other: Any?): Boolean {
-        if (other === this) return true
-        if (other !is Map<*, *>) return false
-        if (other.size != size) return false
-
-        return other.entries.all { (key, value) -> value == map[key] }
-    }
-
-    override fun hashCode(): Int = entries.hashCode()
-
-    override fun toString(): String = entries.joinToString(", ", "{", "}")
 
     private fun wrap(newMap: Map<K, @UnsafeVariance V>): PersistentMap<K, V> = when {
         newMap === map -> this